Changed gulp implementation by separating tasks into individual files and all shared variables into one config file. gulp/index.js iterates over all .js files in gulp/tasks and requires them for gulp
