+++ gulp/tasks/build.js
@@ -1,0 +1,13 @@
+/*!
+ * Facebook React Starter Kit | https://github.com/kriasoft/react-starter-kit
+ * Copyright (c) KriaSoft, LLC. All rights reserved. See LICENSE.txt
+ */
+
+var gulp = require('gulp');
+var runSequence = require('run-sequence');
+
+
+// Build the app from source code
+gulp.task('build', ['clean'], function(cb) {
+  runSequence(['vendor', 'assets', 'styles', 'bundle'], cb);
+});
