+++ gulp/tasks/clean.js
@@ -1,0 +1,12 @@
+/*!
+ * Facebook React Starter Kit | https://github.com/kriasoft/react-starter-kit
+ * Copyright (c) KriaSoft, LLC. All rights reserved. See LICENSE.txt
+ */
+
+var gulp = require('gulp');
+var del = require('del');
+
+// Clean output directory
+gulp.task('clean', del.bind(
+  null, ['.tmp', 'build/*', '!build/.git'], {dot: true}
+));
