+++ gulp/tasks/default.js
@@ -1,0 +1,9 @@
+/*!
+ * Facebook React Starter Kit | https://github.com/kriasoft/react-starter-kit
+ * Copyright (c) KriaSoft, LLC. All rights reserved. See LICENSE.txt
+ */
+
+var gulp = require('gulp');
+
+// The default task
+gulp.task('default', ['sync']);
