+++ gulp/tasks/serve.js
@@ -1,0 +1,43 @@
+/*!
+ * Facebook React Starter Kit | https://github.com/kriasoft/react-starter-kit
+ * Copyright (c) KriaSoft, LLC. All rights reserved. See LICENSE.txt
+ */
+
+var gulp = require('gulp');
+var $ = require('gulp-load-plugins')();
+var config = require('../config');
+
+// Launch a Node.js/Express server
+gulp.task('serve', ['build:watch'], function(cb) {
+
+  var started = false;
+  var cp = require('child_process');
+  var assign = require('react/lib/Object.assign');
+
+  var server = (function startup() {
+    var child = cp.fork(config.server.dest, {
+      env: assign({NODE_ENV: 'development'}, process.env)
+    });
+    child.once('message', function(message) {
+      if (message.match(/^online$/)) {
+        if (config.browserSync) {
+          config.browserSync.reload();
+        }
+        if (!started) {
+          started = true;
+          gulp.watch(config.server.src, function() {
+            $.util.log('Restarting development server.');
+            server.kill('SIGTERM');
+            server = startup();
+          });
+          cb();
+        }
+      }
+    });
+    return child;
+  })();
+
+  process.on('exit', function() {
+    server.kill('SIGTERM');
+  });
+});
